move footer to elements
