+++ src/App.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useContext } from 'react'
 import './App.css'
-import Footer from './components/Footer'
-import Header from './components/Header'
+import { Footer } from 'src/components/Layout'
+import Header from 'src/components/Header'
 import BookmarksSidebar from './bookmark/BookmarksSidebar'
 import { MarketingBanner } from 'src/components/Elements/MarketingBanner'
 import ScrollCardsNavigator from './components/ScrollCardsNavigator'
